extract banner to separate component
